+++ templates/layout.php
@@ -1,12 +1,3 @@
-<?php
-
-$is_auth = (bool) rand(0, 1);
-
-$user_name = 'Константин';
-$user_avatar = 'img/user.jpg';
-
-?>
-
 <!DOCTYPE html>
 <html lang="ru">
 <head>
@@ -62,11 +53,6 @@ $user_avatar = 'img/user.jpg';
 <footer class="main-footer">
     <nav class="nav">
         <ul class="nav__list container">
-            <?php
-                $categories = ["Доски и лыжи", "Крепления", "Ботинки", "Одежда", "Инструменты", "Разное"];
-                $cat_count = count($categories);
-                $index = 0;
-            ?>
             <?php while ($index < $cat_count): ?>
                 <li class="nav__item">
                     <a href="all-lots.html"><?= $categories[$index]; ?></a>
